Add dynamic_step.hpp for base templates declaration
Replace all scattered around forward declarations of
`dynamic_{x,y,xy}_step_type` with this common header,
where the templates are also documented.

Refine documentation of `transposed_type` - this is kept
in locator.hpp as it is not just related to the dynamic step.
