+++ include/boost/gil/extension/dynamic_image/image_view_factory.hpp
@@ -10,6 +10,7 @@
 
 #include <boost/gil/extension/dynamic_image/any_image_view.hpp>
 
+#include <boost/gil/dynamic_step.hpp>
 #include <boost/gil/image_view_factory.hpp>
 #include <boost/gil/point.hpp>
